+++ src/main/java/de/markusbordihn/easynpc/client/screen/editor/DialogEditorScreen.java
@@ -25,6 +25,7 @@ import de.markusbordihn.easynpc.Constants;
 import de.markusbordihn.easynpc.client.screen.components.AddButton;
 import de.markusbordihn.easynpc.client.screen.components.CancelButton;
 import de.markusbordihn.easynpc.client.screen.components.Checkbox;
+import de.markusbordihn.easynpc.client.screen.components.CloseButton;
 import de.markusbordihn.easynpc.client.screen.components.DeleteButton;
 import de.markusbordihn.easynpc.client.screen.components.DialogButton;
 import de.markusbordihn.easynpc.client.screen.components.DialogButtonButton;
@@ -46,7 +47,6 @@ import java.util.Set;
 import java.util.UUID;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.components.Button;
-import net.minecraft.client.gui.components.ImageButton;
 import net.minecraft.client.gui.screens.ConfirmScreen;
 import net.minecraft.client.gui.screens.inventory.AbstractContainerScreen;
 import net.minecraft.client.multiplayer.ClientLevel;
@@ -112,6 +112,7 @@ public class DialogEditorScreen extends AbstractContainerScreen<DialogEditorMenu
   private String dialogTextValue = "";
   private boolean dialogTranslateValue = false;
 
+  @OnlyIn(Dist.CLIENT)
   public DialogEditorScreen(DialogEditorMenu menu, Inventory inventory, Component component) {
     super(menu, inventory, component);
 
@@ -244,6 +245,11 @@ public class DialogEditorScreen extends AbstractContainerScreen<DialogEditorMenu
     this.rightPos = this.leftPos + this.imageWidth;
     this.bottomPos = this.topPos + this.imageHeight;
 
+    // Allow repeated key events
+    if (this.minecraft != null) {
+      this.minecraft.keyboardHandler.setSendRepeatsToGui(true);
+    }
+
     // Home Button
     this.homeButton =
         this.addRenderableWidget(
@@ -344,7 +350,7 @@ public class DialogEditorScreen extends AbstractContainerScreen<DialogEditorMenu
         buttonTopPos += 20;
         buttonLeftPos = buttonBaseLeftPos;
       }
-      Button dialogButton =
+      Button dialogActionButton =
           new DialogButtonButton(
               buttonLeftPos,
               buttonTopPos,
@@ -355,7 +361,7 @@ public class DialogEditorScreen extends AbstractContainerScreen<DialogEditorMenu
                 NetworkMessageHandler.openDialogButtonEditor(
                     uuid, this.dialogId, dialogButtonData.getId(), formerConfigurationType);
               });
-      this.addRenderableWidget(dialogButton);
+      this.addRenderableWidget(dialogActionButton);
       buttonLeftPos += buttonWidth + buttonSpace;
       buttonIndex++;
     }
@@ -382,15 +388,7 @@ public class DialogEditorScreen extends AbstractContainerScreen<DialogEditorMenu
     // Close Button
     this.closeButton =
         this.addRenderableWidget(
-            new ImageButton(
-                this.rightPos - 15,
-                this.topPos + 6,
-                10,
-                10,
-                64,
-                38,
-                Constants.TEXTURE_CONFIGURATION,
-                onPress -> closeScreen()));
+            new CloseButton(this.rightPos - 15, this.topPos + 4, onPress -> closeScreen()));
 
     // Save Button
     this.saveButton =
@@ -443,6 +441,14 @@ public class DialogEditorScreen extends AbstractContainerScreen<DialogEditorMenu
     if (this.dialogLabelCheckbox != null && this.dialogNameToLabelButton != null) {
       this.dialogNameToLabelButton.active = !this.dialogLabelCheckbox.selected();
     }
+  }
+
+  @Override
+  public void removed() {
+    if (this.minecraft != null) {
+      this.minecraft.keyboardHandler.setSendRepeatsToGui(false);
+    }
+    super.removed();
   }
 
   @Override