+++ src/main/java/de/markusbordihn/easynpc/entity/EasyNPCEntityData.java
@@ -35,11 +35,9 @@ import de.markusbordihn.easynpc.entity.data.EntityProfessionData;
 import de.markusbordihn.easynpc.entity.data.EntityScaleData;
 import de.markusbordihn.easynpc.entity.data.EntitySkinData;
 import de.markusbordihn.easynpc.entity.data.EntityTradingData;
-import de.markusbordihn.easynpc.entity.npc.Humanoid.Variant;
 import de.markusbordihn.easynpc.utils.TextUtils;
 import java.util.List;
 import java.util.UUID;
-import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
@@ -52,17 +50,27 @@ import net.minecraft.sounds.SoundEvent;
 import net.minecraft.sounds.SoundEvents;
 import net.minecraft.util.TimeUtil;
 import net.minecraft.util.valueproviders.UniformInt;
+import net.minecraft.world.SimpleContainer;
 import net.minecraft.world.entity.AgeableMob;
 import net.minecraft.world.entity.EntitySelector;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ExperienceOrb;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.NeutralMob;
 import net.minecraft.world.entity.Pose;
 import net.minecraft.world.entity.ai.goal.GoalSelector;
 import net.minecraft.world.entity.ai.navigation.GroundPathNavigation;
+import net.minecraft.world.entity.monster.CrossbowAttackMob;
+import net.minecraft.world.entity.monster.RangedAttackMob;
+import net.minecraft.world.entity.npc.InventoryCarrier;
 import net.minecraft.world.entity.npc.Npc;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.entity.projectile.Projectile;
+import net.minecraft.world.item.BowItem;
+import net.minecraft.world.item.CrossbowItem;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
+import net.minecraft.world.item.ProjectileWeaponItem;
 import net.minecraft.world.item.trading.Merchant;
 import net.minecraft.world.item.trading.MerchantOffer;
 import net.minecraft.world.item.trading.MerchantOffers;
@@ -82,8 +90,11 @@ public class EasyNPCEntityData extends AgeableMob
         EntityScaleData,
         EntitySkinData,
         EntityTradingData,
+        RangedAttackMob,
+        CrossbowAttackMob,
         Npc,
         Merchant,
+        InventoryCarrier,
         NeutralMob {
 
   protected static final Logger log = LogManager.getLogger(Constants.LOG_NAME);
@@ -92,21 +103,23 @@ public class EasyNPCEntityData extends AgeableMob
       SynchedEntityData.defineId(EasyNPCEntityData.class, EntityDataSerializers.BOOLEAN);
   private static final EntityDataAccessor<String> DATA_VARIANT =
       SynchedEntityData.defineId(EasyNPCEntityData.class, EntityDataSerializers.STRING);
-  // Stored Entity Data Tags
+  private static final String DATA_INVENTORY_TAG = "Inventory";
   private static final String DATA_POSE_TAG = "Pose";
   private static final String DATA_TAME_TAG = "Tame";
   private static final String DATA_VARIANT_TAG = "Variant";
+  private static final String DATA_EASY_NPC_DATA_VERSION_TAG = "EasyNPCVersion";
+  private static final int NPC_DATA_VERSION = 1;
   private static final UniformInt PERSISTENT_ANGER_TIME = TimeUtil.rangeOfSeconds(20, 39);
-  // Entity Data
   private final CustomEntityData customEntityData = new CustomEntityData(this);
+  private final SimpleContainer inventory = new SimpleContainer(8);
   protected MerchantOffers offers;
   protected int attackAnimationTick;
-  // Cache
   private boolean syncedDataLoaded = false;
   private boolean isPreview = false;
   private Player tradingPlayer;
   private int remainingPersistentAngerTime;
   private UUID persistentAngerTarget;
+  private int npcDataVersion = -1;
 
   public EasyNPCEntityData(EntityType<? extends EasyNPCEntity> entityType, Level level) {
     super(entityType, level);
@@ -185,7 +198,7 @@ public class EasyNPCEntityData extends AgeableMob
   }
 
   public void notifyTradeUpdated(ItemStack itemStack) {
-    if (!this.level.isClientSide && this.ambientSoundTime > -this.getAmbientSoundInterval() + 20) {
+    if (!this.isClientSide() && this.ambientSoundTime > -this.getAmbientSoundInterval() + 20) {
       this.ambientSoundTime = -this.getAmbientSoundInterval();
       this.playSound(
           this.getTradeUpdatedSound(!itemStack.isEmpty()),
@@ -315,11 +328,11 @@ public class EasyNPCEntityData extends AgeableMob
     this.deserializeNBT(compoundTag);
   }
 
-  public List<Player> getPlayersInRange(Double range) {
+  public List<? extends Player> getPlayersInRange(Double range) {
     return this.level.players().stream()
         .filter(EntitySelector.NO_SPECTATORS)
         .filter(entity -> this.closerThan(entity, range))
-        .collect(Collectors.toList());
+        .toList();
   }
 
   protected void defineCustomData() {
@@ -332,12 +345,8 @@ public class EasyNPCEntityData extends AgeableMob
     return this.syncedDataLoaded;
   }
 
-  public boolean isClientSideExecuted() {
-    return this.level.isClientSide;
-  }
-
-  public boolean isServerSideExecuted() {
-    return !this.level.isClientSide;
+  public int getNPCDataVersion() {
+    return this.npcDataVersion;
   }
 
   @Override
@@ -435,6 +444,11 @@ public class EasyNPCEntityData extends AgeableMob
   @Override
   public void addAdditionalSaveData(CompoundTag compoundTag) {
     super.addAdditionalSaveData(compoundTag);
+
+    // Add version tag.
+    compoundTag.putInt(DATA_EASY_NPC_DATA_VERSION_TAG, NPC_DATA_VERSION);
+
+    // Add additional data.
     this.addAdditionalActionData(compoundTag);
     this.addAdditionalAttackData(compoundTag);
     this.addAdditionalAttributeData(compoundTag);
@@ -457,16 +471,35 @@ public class EasyNPCEntityData extends AgeableMob
       compoundTag.putString(DATA_VARIANT_TAG, this.getVariant().name());
     }
 
+    // Inventory
+    if (!this.inventory.isEmpty()) {
+      compoundTag.put(DATA_INVENTORY_TAG, this.inventory.createTag());
+    }
+
     // Handle tame state.
     compoundTag.putBoolean(DATA_TAME_TAG, this.isTame());
-
-    // Handle anger target.
-    this.addPersistentAngerSaveData(compoundTag);
   }
 
   @Override
   public void readAdditionalSaveData(CompoundTag compoundTag) {
     super.readAdditionalSaveData(compoundTag);
+
+    // Read version tag.
+    if (compoundTag.contains(DATA_EASY_NPC_DATA_VERSION_TAG)) {
+      this.npcDataVersion = compoundTag.getInt(DATA_EASY_NPC_DATA_VERSION_TAG);
+      if (this.npcDataVersion > NPC_DATA_VERSION) {
+        log.warn("Incompatible Easy NPC Data with version {} for {}!", this.npcDataVersion, this);
+      } else if (this.npcDataVersion < NPC_DATA_VERSION) {
+        log.warn(
+            "Outdated Easy NPC Data with version {} for {}!Will try to convert data to new format.",
+            this.npcDataVersion,
+            this);
+      }
+    } else {
+      log.warn("Legacy Easy NPC Data for {}! Will try to convert data to new format.", this);
+    }
+
+    // Read additional data.
     this.readAdditionalActionData(compoundTag);
     this.readAdditionalAttackData(compoundTag);
     this.readAdditionalAttributeData(compoundTag);
@@ -499,8 +532,10 @@ public class EasyNPCEntityData extends AgeableMob
       this.setTame(compoundTag.getBoolean(DATA_TAME_TAG));
     }
 
-    // Read persistent anger target.
-    this.readPersistentAngerSaveData(this.level, compoundTag);
+    // Inventory
+    if (compoundTag.contains(DATA_INVENTORY_TAG, 9)) {
+      this.inventory.fromTag(compoundTag.getList(DATA_INVENTORY_TAG, 10));
+    }
 
     // Register attribute based objectives
     this.registerAttributeBasedObjectives();
@@ -548,8 +583,8 @@ public class EasyNPCEntityData extends AgeableMob
   }
 
   @Override
-  public void setRemainingPersistentAngerTime(int p_21673_) {
-    this.remainingPersistentAngerTime = p_21673_;
+  public void setRemainingPersistentAngerTime(int remainingPersistentAngerTime) {
+    this.remainingPersistentAngerTime = remainingPersistentAngerTime;
   }
 
   @Nullable
@@ -559,14 +594,58 @@ public class EasyNPCEntityData extends AgeableMob
   }
 
   @Override
-  public void setPersistentAngerTarget(@org.jetbrains.annotations.Nullable UUID p_21672_) {
-    this.persistentAngerTarget = p_21672_;
+  public void setPersistentAngerTarget(UUID targetUUID) {
+    this.persistentAngerTarget = targetUUID;
   }
 
   @Override
   public void startPersistentAngerTimer() {
     this.setRemainingPersistentAngerTime(PERSISTENT_ANGER_TIME.sample(this.random));
-    this.setRemainingPersistentAngerTime(PERSISTENT_ANGER_TIME.sample(this.random));
+  }
+
+  @Override
+  public void setChargingCrossbow(boolean isCharging) {
+    this.entityData.set(IS_CHARGING_CROSSBOW, isCharging);
+  }
+
+  @Override
+  public void performRangedAttack(LivingEntity livingEntity, float damage) {
+    if (this.getMainHandItem().getItem() instanceof BowItem) {
+      this.performBowAttack(livingEntity, damage);
+    } else if (this.getMainHandItem().getItem() instanceof CrossbowItem) {
+      EntityAttackData.addChargedProjectile(this.getMainHandItem(), new ItemStack(Items.ARROW, 1));
+      this.performCrossbowAttack(this, 1.6F);
+    }
+  }
+
+  @Override
+  public void shootCrossbowProjectile(
+      LivingEntity livingEntity, ItemStack itemStack, Projectile projectile, float rangeFactor) {
+    log.info("Shoot crossbow {} {} {}", itemStack, projectile, rangeFactor);
+    this.shootCrossbowProjectile(this, livingEntity, projectile, rangeFactor, 1.6F);
+  }
+
+  @Override
+  public void onCrossbowAttackPerformed() {
+    this.noActionTime = 0;
+  }
+
+  @Override
+  public boolean canFireProjectileWeapon(ProjectileWeaponItem projectileWeaponItem) {
+    return projectileWeaponItem == Items.CROSSBOW || projectileWeaponItem == Items.BOW;
+  }
+
+  @Override
+  public SimpleContainer getInventory() {
+    return this.inventory;
+  }
+
+  protected ItemStack addToInventory(ItemStack itemStack) {
+    return this.inventory.addItem(itemStack);
+  }
+
+  protected boolean canAddToInventory(ItemStack itemStack) {
+    return this.inventory.canAddItem(itemStack);
   }
 
   // Default Variants